LPD-41517 Update CKE4 tests and add few simple tests for CKE5
